+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## [1.0.0] 20xx-xx-xx
 ### Original Release
-- Started project from [Tailwind Starter Kit by Creative Tim](https://www.creative-tim.com/learning-lab/tailwind-starter-kit#/presentation)
+- Started project from [Tailwind Starter Kit by Creative Tim](https://www.creative-tim.com/learning-lab/tailwind-starter-kit/presentation?ref=twnjs-changelog)
 - Added design from Tailwind Starter Kit by Creative Tim